Add files via upload
This python script can be used to collect all historical weather data from visual crossing for a single location. This file is developed to collect all the historical weather data for Washington State. This script will required the base paid API key. Additionally if this script is use the user will need to add their API Key and insert the proper file paths as necessary.