Daily Weather Extraction Module
This python script will allow the user to have the current day weather added to the data repository CSV that was collected after running the Historical Weather Module. This will keep all data up to date and allow the model to look at real time data. Additionally if this script is used the user will need to add their API Key and insert the proper file paths as necessary.