La Nina weather forecast model
This final code script is the forecast model that was developed for after the data collection steps had been completed. This model use three algorithms from SKLearn python library; Isolation Forest, Local Outlier Factor, and Support Vector Machine. This model was trying attempting to accurately predict if the La Nina weather cycle was occurring on a given day based on the data available from visual crossing. This model produced an accuracy of 35% or that the model could only correctly predict the La Nina weather cycle 35% of the time. Additionally if this script is use the user will need to add their API Key and insert the proper file paths as necessary.